feat: demo with file backend instead of templates
