+++ demo/data/.internal/accounts/index/owner/6bf4fe03-20c1-419d-9934-2b7533296edf$.json
@@ -1,0 +1,1 @@
+{"key":"accounts/index/owner/6bf4fe03-20c1-419d-9934-2b7533296edf","payload":["c5b28411-2340-4820-8f4f-62c209c20172"]}
