+++ demo/data/.internal/accounts/index/password/email/catalog@example.org$.json
@@ -1,0 +1,1 @@
+{"key":"accounts/index/password/email/catalog%40example.org","payload":["c5b28411-2340-4820-8f4f-62c209c20172"]}
