+++ demo/data/.internal/accounts/index/password/email/demo@example.org$.json
@@ -1,0 +1,1 @@
+{"key":"accounts/index/password/email/demo%40example.org","payload":["d3156f11-ffb2-42f3-b928-b9752a9873ce"]}
