+++ demo/data/.internal/accounts/index/password/email/ruben@example.org$.json
@@ -1,0 +1,1 @@
+{"key":"accounts/index/password/email/ruben%40example.org","payload":["f644f883-ef0f-4986-b5ff-df6866707cf6"]}
