+++ demo/data/.internal/setup/current-base-url$.json
@@ -1,0 +1,1 @@
+{"key":"setup/current-base-url","payload":"http://localhost:3000/"}
