+++ demo/data/.internal/setup/current-server-version$.json
@@ -1,0 +1,1 @@
+{"key":"setup/current-server-version","payload":"7.0.2"}
