+++ demo/data/.internal/setup/rootInitialized$.json
@@ -1,0 +1,1 @@
+{"key":"setup/rootInitialized","payload":true}
