+++ demo/data/.internal/setup/v6-migration$.json
@@ -1,0 +1,1 @@
+{"key":"setup/v6-migration","payload":true}
