+++ demo/data/catalog/.meta
@@ -1,0 +1,1 @@
+<http://localhost:3000/catalog/> a <http://www.w3.org/ns/pim/space#Storage>.
