+++ demo/data/demo/.meta
@@ -1,0 +1,1 @@
+<http://localhost:3000/demo/> a <http://www.w3.org/ns/pim/space#Storage>.
