+++ demo/data/ruben/private/.meta
@@ -1,0 +1,13 @@
+@prefix derived: <urn:npm:solid:derived-resources:> .
+
+<> derived:derivedResource [
+        derived:template "derived/bday";
+        derived:selector <./data>;
+        derived:filter <http://localhost:3000/catalog/public/filters/bday>
+    ].
+    
+<> derived:derivedResource [
+        derived:template "derived/age";
+        derived:selector <./data>;
+        derived:filter <http://localhost:3000/catalog/public/filters/age>
+    ].
