+++ demo/data/ruben/profile/card$.ttl
@@ -1,0 +1,13 @@
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+@prefix foaf: <http://xmlns.com/foaf/0.1/>.
+@prefix solid: <http://www.w3.org/ns/solid/terms#>.
+
+<> a foaf:PersonalProfileDocument;
+    foaf:maker <http://localhost:3000/ruben/profile/card#me>;
+    foaf:primaryTopic <http://localhost:3000/ruben/profile/card#me>.
+
+<http://localhost:3000/ruben/profile/card#me> a foaf:Person ;
+    foaf:name  "Ruben Verborgh"@en, "Ruben Verborgh"@nl;
+    rdfs:label "Ruben Verborgh"@en, "Ruben Verborgh"@nl;
+    solid:umaServer "http://localhost:4000/uma" ;
+    solid:oidcIssuer <http://localhost:3000/> .
