+++ package.json
@@ -56,13 +56,14 @@
     "build:all": "yarn workspaces foreach --exclude . -A -pi -t run build",
     "test:all": "yarn workspaces foreach --exclude . -A -pi run test",
     "start:all": "yarn workspaces foreach --exclude . -A -pi run start",
+    "start:demo": "yarn workspaces foreach --exclude . -A -pi run demo",
+    "script:demo": "yarn exec ts-node ./scripts/test-demo.ts",
     "script:public": "yarn exec ts-node ./scripts/test-public.ts",
     "script:private": "yarn exec ts-node ./scripts/test-private.ts",
-    "script:derived": "yarn exec ts-node ./scripts/test-derived.ts",
     "script:registration": "yarn exec ts-node ./scripts/test-registration.ts",
     "script:ucp-enforcement": "yarn exec ts-node ./scripts/test-ucp-enforcement.ts",
     "script:uma-ucp": "yarn exec ts-node ./scripts/test-uma-ucp.ts",
-    "script:flow": "yarn run script:public && yarn run script:private && yarn run script:derived && yarn run script:uma-ucp && yarn run script:registration && yarn run script:ucp-enforcement"
+    "script:flow": "yarn run script:public && yarn run script:private && yarn run script:uma-ucp && yarn run script:registration && yarn run script:ucp-enforcement"
   },
   "devDependencies": {
     "@commitlint/cli": "^16.1.0",
@@ -77,7 +78,7 @@
     "eslint": "^8.10.0",
     "jest": "^29.7.0",
     "koreografeye": "^0.4.8",
-    "shx": "^0.3.3",
+    "shx": "^0.3.4",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.2.2"
