+++ packages/css/package.json
@@ -60,7 +60,10 @@
     "build:ts": "yarn run -T tsc",
     "build:components": "yarn run -T componentsjs-generator -r uma-css -s src/ -c dist/components -i .componentsignore --lenient",
     "test": "yarn run -T jest --coverage",
-    "start": "yarn run community-solid-server -m . -c ./config/default.json ./config/derived.json ./config/demo.json --seedConfig ./config/seed.json"
+    "start": "yarn run community-solid-server -m . -c ./config/default.json --seedConfig ./config/seed.json",
+    "demo": "yarn run demo:setup && yarn run demo:start",
+    "demo:setup": "yarn run -T shx rm -rf ./tmp && yarn run -T shx cp -R ../../demo/data ./tmp",
+    "demo:start": "yarn run community-solid-server -m . -c ./config/default.json ./config/derived.json -f ./tmp"
   },
   "dependencies": {
     "@solid/community-server": "^7.0.2",
