--- packages/css/templates/pod/base/.meta
@@ -1,3 +1,0 @@
-@prefix pim: <http://www.w3.org/ns/pim/space#>.
-
-<> a pim:Storage.
