--- packages/css/templates/pod/base/README$.md.hbs
@@ -1,27 +1,0 @@
-# Welcome to your pod
-
-## A place to store your data
-Your pod is a **secure storage space** for your documents and data.
-<br>
-You can choose to share those with other people and apps.
-
-As the owner of this pod,
-identified by <a href="{{webId}}">{{webId}}</a>,
-you have access to all of your documents.
-
-## Working with your pod
-The easiest way to interact with pods
-is through Solid apps.
-<br>
-For example,
-you can open your pod in [Databrowser](https://solidos.github.io/mashlib/dist/browse.html?uri={{base.path}}).
-
-## Accessing your account
-To keep track of your pods, webIDs and any other resources,
-you can [log in]({{oidcIssuer}}.account/) to your account.
-There you can, for example, update the owners of this pod.
-
-## Learn more
-The [Solid website](https://solidproject.org/)
-and the people on its [forum](https://forum.solidproject.org/)
-will be glad to help you on your journey.
