--- packages/css/templates/pod/base/private/.meta
@@ -1,22 +1,0 @@
-@prefix derived: <urn:npm:solid:derived-resources:> .
-
-<>
-    derived:derivedResource [
-        derived:template "derived/bday";
-        derived:selector <./data>;
-        derived:filter <../public/filters/bday>
-    ].
-    
-<>
-    derived:derivedResource [
-        derived:template "derived/age";
-        derived:selector <./data>;
-        derived:filter <../public/filters/age>
-    ].
-
-# <>
-#     derived:derivedResource [
-#         derived:template "query{?var}";
-#         derived:selector <selectors/data>;
-#         derived:filter <filters/var>
-#     ].
