--- packages/css/templates/pod/base/profile/card$.ttl.hbs
@@ -1,13 +1,0 @@
-@prefix foaf: <http://xmlns.com/foaf/0.1/>.
-@prefix solid: <http://www.w3.org/ns/solid/terms#>.
-
-<>
-    a foaf:PersonalProfileDocument;
-    foaf:maker <{{webId}}>;
-    foaf:primaryTopic <{{webId}}>.
-
-<{{webId}}>
-    {{#if name}}foaf:name "{{name}}";{{/if}}
-    {{#if umaServer}}solid:umaServer "{{umaServer}}";{{/if}}
-    {{#if oidcIssuer}}solid:oidcIssuer <{{oidcIssuer}}>;{{/if}}
-    a foaf:Person.
