--- packages/css/templates/pod/base/public/.meta
@@ -1,25 +1,0 @@
-#
-# THIS IS A COPY OF ../private/.meta
-
-@prefix derived: <urn:npm:solid:derived-resources:> .
-
-<>
-    derived:derivedResource [
-        derived:template "derived/bday";
-        derived:selector <./data>;
-        derived:filter <filters/bday>
-    ].
-    
-<>
-    derived:derivedResource [
-        derived:template "derived/age";
-        derived:selector <./data>;
-        derived:filter <filters/age>
-    ].
-
-# <>
-#     derived:derivedResource [
-#         derived:template "query{?var}";
-#         derived:selector <selectors/data>;
-#         derived:filter <filters/var>
-#     ].
