--- packages/css/templates/pod/base/public/data$.ttl
@@ -1,23 +1,0 @@
-#
-# THIS IS A COPY OF ../private/data
-
-@prefix ruben: <https://ruben.verborgh.org/profile/#me>.
-@prefix con: <http://www.w3.org/2000/10/swap/pim/contact#>.
-@prefix dbo: <http://dbpedia.org/ontology/>.
-@prefix dbp: <http://dbpedia.org/resource/>.
-@prefix foaf: <http://xmlns.com/foaf/0.1/>.
-@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#>.
-@prefix vcard: <http://www.w3.org/2006/vcard/ns#>.
-@prefix xsd: <http://www.w3.org/2001/XMLSchema#>.
-
-ruben: a foaf:Person;
-  foaf:name  "Ruben Verborgh"@en, "Ruben Verborgh"@nl;
-  rdfs:label "Ruben Verborgh"@en, "Ruben Verborgh"@nl;
-  vcard:fn   "Ruben Verborgh"@en, "Ruben Verborgh"@nl;
-  con:preferredURI "https://ruben.verborgh.org/profile/#me";
-  foaf:givenName "Ruben"@en, "Ruben"@nl;
-  foaf:familyName "Verborgh"@en, "Verborgh"@nl;
-  rdfs:label "Ruben Verborgh"@en, "Ruben Verborgh"@nl;
-  dbo:birthPlace dbp:Ostend;
-  dbo:birthDate "1987-02-28"^^xsd:date;
-  foaf:gender "male"@en.
